Add `GYTxBuilderMonad` for custom build tx impls
