+++ atlas-cardano.cabal
@@ -25,6 +25,7 @@ common common
   default-language:   GHC2021
   default-extensions:
     DataKinds
+    DefaultSignatures
     DeriveAnyClass
     DerivingStrategies
     DerivingVia
